# Improve performance of parsing large texts.