The current implementation of `process_chunk` returns either a text or an array, which is then "flattened" out later, this is a slow process if the input string is large.

This commit makes the chunking method consistently return a string, which also does not need to be flattened out.