@drudru 2015-05-16
Hi ketan - can you provide some benchmarks? I'm all for making the code more performant. Thanks!