@ketan 2015-05-18
In case you're curious what and why we're doing this, I'm attempting to use your library to perform console log coloring in the browser for gocd/gocd#142. The console log can be fairly large leading into several megs.

I've provided a set of benchmark code over on this gist - https://gist.github.com/ketan/da0ce4f6a6a1728b9e02

Here's a set of benchmark numbers -

```
$ INPUT_SIZE_MULTIPLIER=10 node ansi-benchmark.js; INPUT_SIZE_MULTIPLIER=20 node ansi-benchmark.js; INPUT_SIZE_MULTIPLIER=30 node ansi-benchmark.js

Running with input size 10
new-implementation x 25.61 ops/sec ±3.72% (47 runs sampled)
old-implementation x 4.17 ops/sec ±3.74% (15 runs sampled)
Running with input size 20
new-implementation x 11.23 ops/sec ±6.13% (30 runs sampled)
old-implementation x 1.09 ops/sec ±4.57% (7 runs sampled)
Running with input size 30
new-implementation x 8.35 ops/sec ±4.92% (26 runs sampled)
old-implementation x 0.28 ops/sec ±5.23% (5 runs sampled)
```