@ketan 2015-05-20
I have a few more improvements (not performance) that I can make once this is merged.

Thanks for your great work.