@drudru 2015-05-20
Hi - no problem. Here is an update. I updated my node and npms yesterday. Now all the tests are timing out at 1ms. This is crazy. I'm not sure what the culprit is (virtual box, ubuntu, new versions etc.). Once I figure that out, I can move forward on this.