@ketan 2015-05-20
I noticed the tests being flaky about timings; but they run fine most of the time.

I was not running them on a VM though.