@drudru 2015-05-21
ok - pushed up my fixes. Send in your new patch and I'll review it. Thanks!